pyproject.toml
[tool.poetry]
name = "store-server"
version = "0.1.0"
description = ""
authors = ["Rom1xa <abakumoff.roman2018@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.15"
pillow = "^10.4.0"
ipython = "^8.27.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
django-allauth = "^64.2.1"
requests = "^2.32.3"
icecream = "^2.1.3"
coverage = "^7.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
